added example eqtl track
