--- data/_constants.ts
@@ -1,5 +1,0 @@
-// export const FEATURE_SEARCH_ENDPOINT="https://www.niagads.org/genomics/service/track/feature?id="
-export const WEBSERVICE_URL=""
-export const FEATURE_SEARCH_ENDPOINT="/service/track/feature?id="
-export const DEFAULT_FLANK=1000
-export const ALWAYS_ON_TRACKS = ["ideogram", "ruler", "sequence", "ENSEMBL_GENE"];
