+++ tsconfig.json
@@ -24,7 +24,7 @@
       "@components/*": ["./src/components/*"],
       "@utils/*": ["./src/utils/*"],
       "@browser-types/*": ["./src/types/*"],
-      "@data/*": ["./data/*"],
+      "@data/*": ["./src/data/*"],
       "@readers/*": ["./src/readers/*"],
       "@tracks/*": ["./src/tracks/*"],
       "@parsers/*": ["./src/parsers"]
